Last testing with the crawler before the alpha
I did a last few tests before moving to main. I made a simple path for the monster and it worked. I also fixed a bug in the CrawlerBehavior where its path would be null. I also changed the prefab for every room to put the colliders on the wall layer.