Monsters for bata, first pass
Made the Crawler into a prefab, added the pause behavior for it when it hits the player.
Added the Shadow, chases after player constantly, after it chases the player for  a bit it will teleport away.
No interaction with monsters and light is implemented like.
Added an automatic node connector.